Refactor code to support AspNet classic
